requirements.txt
ultralytics
motrackers
numpy
matplotlib
opencv-python
pillow
pyyaml
requests
scipy
torch
torchvision
tqdm
pandas
psutil
